refactor demo scripts and add integration test
