+++ evm/contracts/bridge/BitcoinBridge.sol.template
