+++ evm/contracts/oracle/BitcoinHeader.sol.template
